chore: improve datatable
