+++ src/app/common/DataTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DataTable';
